Move test functions of anonymized to test helper
