+++ metrics_utility/anonymized_rollups/__init__.py
@@ -1,14 +1,11 @@
 from .anonymized_rollups import (
     anonymize_data,
     anonymize_rollups,
-    compute_anonymized_rollup_from_raw_data,
     create_anonymized_object,
     flatten_json_report,
     hash,
-    load_anonymized_rollup_data,
 )
 from .base_anonymized_rollup import BaseAnonymizedRollup
-from .compute_anonymized_rollup import compute_anonymized_rollup
 from .controller_version_anonymized_rollup import ControllerVersionAnonymizedRollup
 from .credentials_anonymized_rollup import CredentialsAnonymizedRollup
 from .events_modules_anonymized_rollup import EventModulesAnonymizedRollup
@@ -32,11 +29,8 @@ __all__ = [
     'TableMetadataAnonymizedRollup',
     'anonymize_data',
     'anonymize_rollups',
-    'compute_anonymized_rollup',
-    'compute_anonymized_rollup_from_raw_data',
     'create_anonymized_object',
     'flatten_json_report',
     'hash',
-    'load_anonymized_rollup_data',
     'sanitize_json',
 ]
