--- metrics_utility/library/anonymize/__init__.py
@@ -1,8 +1,0 @@
-from .anonymized_rollups_processor import anonymized_rollups_processor
-from .compute_from_raw_data import compute_anonymized_rollup_from_raw_data
-
-
-__all__ = [
-    'anonymized_rollups_processor',
-    'compute_anonymized_rollup_from_raw_data',
-]
