--- metrics_utility/library/anonymize/anonymized_rollups_processor.py
@@ -1,7 +1,0 @@
-# call compute_anonymized_rollup
-
-from metrics_utility.anonymized_rollups.compute_anonymized_rollup import compute_anonymized_rollup
-
-
-def anonymized_rollups_processor(db, salt, since, until):
-    return compute_anonymized_rollup(db, salt, since, until)
