--- metrics_utility/library/anonymize/compute_from_raw_data.py
@@ -1,6 +1,0 @@
-# Expose compute_anonymized_rollup_from_raw_data for library access
-
-from metrics_utility.anonymized_rollups.anonymized_rollups import compute_anonymized_rollup_from_raw_data
-
-
-__all__ = ['compute_anonymized_rollup_from_raw_data']
