+++ metrics_utility/test/test_anonymized_rollups/big_test1/test_big_test_job2.py
@@ -22,7 +22,7 @@ import tempfile
 
 import pandas as pd
 
-from metrics_utility.anonymized_rollups.anonymized_rollups import compute_anonymized_rollup_from_raw_data
+from metrics_utility.test.test_anonymized_rollups.helpers import compute_anonymized_rollup_from_raw_data
 from metrics_utility.test.test_anonymized_rollups.big_test1.credentials import credentials
 from metrics_utility.test.test_anonymized_rollups.big_test1.execution_environments import execution_environments
 from metrics_utility.test.test_anonymized_rollups.big_test1.job2 import events, jobhostsummary, jobs
