+++ metrics_utility/test/test_anonymized_rollups/test_multiple_files.py
@@ -26,7 +26,7 @@ from datetime import datetime
 import pandas as pd
 import pytest
 
-from metrics_utility.anonymized_rollups.anonymized_rollups import compute_anonymized_rollup_from_raw_data
+from metrics_utility.test.test_anonymized_rollups.helpers import compute_anonymized_rollup_from_raw_data
 from metrics_utility.test.test_anonymized_rollups.test_credentials_anonymized_rollup import credentials
 from metrics_utility.test.test_anonymized_rollups.test_events_modules_anonymized_rollups import events
 from metrics_utility.test.test_anonymized_rollups.test_execution_environments_anonymized_rollups import execution_environments
