+++ tools/anonymized_db_perf_data/rollup_performance_test.py
@@ -30,7 +30,7 @@ prepare()
 from django.db import connection  # noqa: E402
 
 # Import rollup computation
-from metrics_utility.anonymized_rollups.anonymized_rollups import (  # noqa: E402
+from metrics_utility.test.test_anonymized_rollups.helpers import (  # noqa: E402
     compute_anonymized_rollup_from_raw_data,
 )
