+++ tools/anonymized_tests/run_no_events.py
@@ -49,7 +49,7 @@ prepare()
 
 from django.db import connection  # noqa: E402
 
-from metrics_utility.anonymized_rollups.anonymized_rollups import (  # noqa: E402
+from metrics_utility.test.test_anonymized_rollups.helpers import (  # noqa: E402
     compute_anonymized_rollup_from_raw_data,
 )
 from metrics_utility.library.collectors.controller import (  # noqa: E402
